Move setters to item.rb
