+++ lib/Game/game.rb
@@ -10,10 +10,6 @@ class Game < Item
     @last_played_at = Date.parse(last_played_at)
   end
 
-  def author=(author)
-    @author = author
-    author.items.push(self)
-  end
 
   def can_be_archived?
     parent = super
